Started Class Diagram
As this is a more complicated project, I've started a class diagram to
map out what is needed for the end result. In the interim, the project
will be locally based and stand alone. The class diagram is intended to
lay out the end goal of a web-based, multi-user, learning engine that is
secure, interactive, and provides useful data analytics.